// nodejs/basic-react-template/src/Models/stream.ts
/**
 * Telemetry Query API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { StreamStatus } from './streamStatus';

/**
 * Describes properties of a stream
 */
export interface Stream { 
    /**
     * The primary identifier of the stream
     */
    streamId?: string;
    /**
     * The display name of the stream
     */
    name?: string;
    /**
     * The topic the stream originates from
     */
    topic?: string;
    /**
     * The time the stream was created at
     */
    createdAt?: Date;
    /**
     * The time the stream was last updated at
     */
    lastUpdate?: Date;
    /**
     * Time of the recording. Optional.
     */
    timeOfRecording?: Date;
    /**
     * The earliest data within the session in unix nanoseconds
     */
    dataStart?: number;
    /**
     * The last data within the session in unix nanoseconds
     */
    dataEnd?: number;
    status?: StreamStatus;
    /**
     * Metadata (extra context) for the stream
     */
    metadata?: { [key: string]: string; };
    /**
     * The stream Ids this session is derived from.
     */
    parents?: Array<string>;
    /**
     * The location of the stream
     */
    location?: string;
    /**
     * The time the stream was soft deleted
     */
    softDeleteAt?: Date;
}