// nodejs/basic-react-template/src/Models/event.ts
/**
 * Telemetry Query API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { EventLevel } from './eventLevel';

/**
 * Describes properties of an event
 */
export interface Event { 
    /**
     * Primary Identifier of the event
     */
    eventId?: string;
    /**
     * Display name of the event
     */
    name?: string;
    /**
     * Description of the event
     */
    description?: string;
    /**
     * Optional field for any user defined custom properties.  This could be a string describing the possible values of this event
     */
    customProperties?: string;
    /**
     * The stream id(s) the event belongs to.  If events are queried for multiple streams, the streams that have this event will have their ids listed here
     */
    streamIds?: Array<string>;
    level?: EventLevel;
    /**
     * Location of the event within the event group hierarchy
     */
    location?: string;
}

export interface QuixEvent {
    topicId: string;
    topicName: string;
    streamId: string;
    timestamp: number;
    tags: Record<string, any>;
    id: string;
    value: string;
  }